Refactor: from class to record, move package to org.fungover.thunder
